warning fixes

+++ Src/Character/charMovement.c
@@ -27,7 +27,7 @@ int check_limits(CP_Vector dir)
 
 	if (dir.x >= 0 && dir.x < getMap()->width && dir.y >= 0 && dir.y < getMap()->height) // check within bounds before checking tile types
 	{
-		char var = getMap()->map_arr[map_get_index(dir.x, dir.y, getMap()->width)];
+		char var = getMap()->map_arr[map_get_index((int)dir.x, (int)dir.y, (int)getMap()->width)];
 		if ((var >= WALL_1 && var <= WALL_17) || (var >= WALL_19 && var <= WALL_20))
 		{
 			return 0;
@@ -114,11 +114,11 @@ void hardware_handler(void)
 int check_Chest(CP_Vector pos)
 {
 	pos.x -= mapOffset[currLvl];
-	char var = getMap()->map_arr[map_get_index(pos.x, pos.y, getMap()->width)];
+	char var = getMap()->map_arr[map_get_index((int)pos.x, (int)pos.y, getMap()->width)];
 
 	if (var == CHEST_1)
 	{
-		getMap()->map_arr[map_get_index(pos.x, pos.y, getMap()->width)] = CHEST_2;
+		getMap()->map_arr[map_get_index((int)pos.x, (int)pos.y, getMap()->width)] = CHEST_2;
 
 		return randomise_all(currLvl);
 	}
